requirements.txt
streamlit>=1.40.0
openai>=1.50.0
pydantic>=2.0.0
python-dotenv>=1.0.0
plotly>=5.24.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
requests>=2.31.0
langchain>=0.3.0
langchain-mistralai>=0.2.0
langchain-core>=0.3.0
guardrails-ai>=0.10.0
